wined3d: Don't synchronize DISCARD maps
Implementing NO_OVERWRITE maps in addition sounds easy at first, but is
fairly tricky because the main thread might map a buffer after
scheduling a command that reads from that buffer. In this situation, the
worker thread receives an error from GL. It would have to wait until the
application unmaps that buffer again. This in turn deadlocks if the
application forces us to wait for the CS before unmapping the buffer,
e.g. with a non-dynamic map of a separate resource.

The solution is GL_ARB_buffer_storage, which allows us to lift the
restriction that a buffer must be unmapped before it can be used for
texture uploads.

For DISCARD maps this is not an issue because we use new or recycled
buffers that are not referenced by pending commands.